Adds function for getting the bounds and renames get_surface
There are several functions with the prefix `get_` which are intended to
be available to clients of the module to use to retrieve information
about a model and/or source. The function `get_surface` selects a subset
of data from the dataset for the vertical layer nearest the surface, so
a more appropriate name would be `_select_surface`.