Adds check to ensure that valid data exists in horizontal grid
If a subsetting bounding box causes the output dataset to have
horizontal dimensions of size 0, it creates a valid dataset but it's
unexpected and undesired. This commit updates the fetching code to make
checks on the subsetted dataset to ensure that some data exists before
attempting to write anything to disk or fetch unnecessary data from the
remote server.